api/serializers.py
from rest_framework import serializers
from .models import Stock, Portfolio


class StockSerializer(serializers.ModelSerializer):

    class Meta:
        model = Stock
        fields = '__all__'


class PortfolioSerializer(serializers.ModelSerializer):

    class Meta:
        model = Portfolio
        fields = '__all__'
        depth = 2
