requirements.txt
Django
djangorestframework
ipython
flake8
